+++ CHANGES.md
@@ -1,6 +1,7 @@
 0.7.0
 -----
 * New timer subsystem! Bullet point-by-bullet point details below
+* Celluloid#after registers a callback to fire after a given time interval
 * Celluloid.sleep and Celluloid#sleep let an actor continue processing messages
 * Celluloid.receive and Celluloid#receive now accept an optional timeout
 * Celluloid::Mailbox#receive now accepts an optional timeout